refactor(flow): consolidate hooks into context
