--- packages/giselle-engine/src/react/flow/hooks/use-properties-panel.ts
@@ -1,13 +1,0 @@
-import { useState } from "react";
-
-export function usePropertiesPanel() {
-	const [propertiesTab, setPropertiesTab] = useState("");
-	const [openPropertiesPanel, setOpenPropertiesPanel] = useState(false);
-
-	return {
-		propertiesTab,
-		setPropertiesTab,
-		openPropertiesPanel,
-		setOpenPropertiesPanel,
-	};
-}
