--- packages/giselle-engine/src/react/flow/hooks/use-workflow-designer.ts
@@ -1,12 +1,0 @@
-import { useContext } from "react";
-import { WorkflowDesignerContext } from "../context";
-
-export function useWorkflowDesigner() {
-	const context = useContext(WorkflowDesignerContext);
-	if (context === undefined) {
-		throw new Error(
-			"useWorkflowDesigner must be used within a WorkflowDesignerProvider",
-		);
-	}
-	return context;
-}
